unit tests and fixes about account backend (#9)
* unit tests and fixes

* removed usedlesses todo

+++ server/resolvers.ts
@@ -7,7 +7,7 @@ import { algoQuery, usersQuery } from "./query";
 const resolvers: Resolvers = {
   Query: {
     searchProducts: async (_obj, _args, _context, _info) => {
-      return algoQuery<Product>("SELECT * FROM products WHERE name LIKE ? OR keywords LIKE ? LIMIT 10", [`%${_args.query}%`, `%${_args.query}%`]);
+      return algoQuery<Product>("SELECT * FROM products WHERE name LIKE ? OR keywords LIKE ? LIMIT 10", [`%${_args.query}%`, `%${_args.query}%`]); // TODO: TEST
     },
     account: async (_obj, _args, context, _info) => {
       if (!context.user)
@@ -43,12 +43,12 @@ const resolvers: Resolvers = {
   },
   Mutation: {
     login: async (_obj, args, _context, _info) => {
-      const loginQuery = await usersQuery("SELECT * FROM users WHERE email = ? AND password = ? LIMIT 1", [args.email, args.passwordSHA256]);
+      const loginQuery = await usersQuery<any>("SELECT * FROM users WHERE email = ? AND password = ? LIMIT 1", [args.email, args.passwordSHA256]);
       if (loginQuery.length != 1)
         throw new AuthenticationError("Invalid email or password");
       const newToken = [...Array(64)].map(() => Math.random().toString(36)[2]).join('');
 
-      await usersQuery("UPDATE users SET token = ?", [newToken]);
+      await usersQuery("UPDATE users SET token = ? WHERE id = ?", [newToken, loginQuery[0].id]);
       return newToken;
     },
     register: async (_obj, args, _context, _info) => {
@@ -66,6 +66,7 @@ const resolvers: Resolvers = {
       if (args.obligations.some(c1 => !Obligations.find(c2 => c1.id == c2.id)))
         throw new UserInputError("Invalid obligation id");
 
+
       await usersQuery("DELETE FROM obligations WHERE userId = ?", [context.user.id]);
       await Promise.all(args.obligations.map(o => usersQuery("INSERT INTO obligations (id, userId) VALUES (?, ?)", [o.id, context.user.id])));
       return true;
@@ -76,6 +77,15 @@ const resolvers: Resolvers = {
       if (args.criterions.some(c1 => !Criterions.find(c2 => c1.id == c2.id)))
         throw new UserInputError("Invalid criterion id");
 
+      if (args.criterions.some(c => c.position === 0))
+        throw new UserInputError("Criterions position must start with 1");
+
+      Array.from({ length: args.criterions.length }).forEach((_e, i) => {
+        if (args.criterions.find(e => e.position === (i + 1)) == undefined)
+          throw new UserInputError("Criterions position are incorrect (not 1, 2, 3...)");
+      })
+
+
       await usersQuery("DELETE FROM criterions WHERE userId = ?", [context.user.id]);
       await Promise.all(args.criterions.map(o => usersQuery("INSERT INTO criterions (id, userId, position) VALUES (?, ?, ?)", [o.id, context.user.id, o.position])));
       return true;
@@ -83,7 +93,7 @@ const resolvers: Resolvers = {
     removeAccount: async (_obj, args, context, _info) => {
       if (!context.user)
         throw new AuthenticationError("please login");
-      const login = await usersQuery("SEELCT * FROM users WHERE id = ? AND password = ? LIMIT 1", [context.user.id, args.passwordSHA256]);
+      const login = await usersQuery("SELECT * FROM users WHERE id = ? AND password = ? LIMIT 1", [context.user.id, args.passwordSHA256]);
       if (login.length != 1)
         throw new AuthenticationError("Invalid password");
       await usersQuery("DELETE FROM users WHERE id = ?", [context.user.id]);
